added functions in Feed and Follow
