+++ anime_post/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
-from anime_post.models import AnimePost
+from anime_post.models import AnimePost, Feed
 
 # Register your models here.
 admin.site.register(AnimePost)
+admin.site.register(Feed)
