+++ anime_user/admin.py
@@ -1,10 +1,13 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
-from anime_user.models import AnimeUser
+from anime_user.models import AnimeUser, Follow
 
 # Register your models here.
 
 admin.site.register(AnimeUser, UserAdmin)
+admin.site.register(Follow)
 
-UserAdmin.fieldsets += ('Custom fields set', {'fields': ('about_me', 'avatar', 'followers')}),
 
+UserAdmin.fieldsets += (
+    ("Custom fields set", {"fields": ("about_me", "avatar", "followers")}),
+)
